move the Classrooms page and its modal to the Classroom folder
