+++ src/pages/Settings/Classrooms/LeaveClassroomModal.sc.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { blue600 } from "../../components/colors";
+import { blue600 } from "../../../components/colors";
 
 const ColorAccent = styled.span`
   font-size: inherit;
